green-compute-oracle/backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
httpx==0.26.0
cryptography==42.0.0
python-multipart==0.0.6
prometheus-client==0.19.0
pytest==8.0.0
